[Pal] Move x86_64-specifics from db_misc.c to db_main-x86_64.c
