+++ Pal/src/host/Linux-SGX/db_main-x86_64.c
@@ -164,3 +164,31 @@ int _DkGetCPUInfo (PAL_CPU_INFO* ci) {
 
     return rv;
 }
+
+size_t _DkRandomBitsRead(void* buffer, size_t size) {
+    uint32_t rand;
+    for (size_t i = 0; i < size; i += sizeof(rand)) {
+        rand = rdrand();
+        memcpy(buffer + i, &rand, MIN(sizeof(rand), size - i));
+    }
+    return 0;
+}
+
+int _DkSegmentRegisterSet(int reg, const void* addr) {
+    /* GS is internally used, denied any access to it */
+    if (reg != PAL_SEGMENT_FS)
+        return -PAL_ERROR_DENIED;
+
+    SET_ENCLAVE_TLS(fsbase, (void*)addr);
+    wrfsbase((uint64_t)addr);
+    return 0;
+}
+
+int _DkSegmentRegisterGet(int reg, void** addr) {
+    /* GS is internally used, denied any access to it */
+    if (reg != PAL_SEGMENT_FS)
+        return -PAL_ERROR_DENIED;
+
+    *addr = (void*)GET_ENCLAVE_TLS(fsbase);
+    return 0;
+}
