+++ Pal/src/host/Linux/db_main-x86_64.c
@@ -7,6 +7,8 @@
  * This file contains x86_64-specific functions of the PAL loader.
  */
 
+#include <asm/prctl.h>
+
 #include "api.h"
 #include "bogomips.h"
 #include "cpu.h"
@@ -151,4 +153,56 @@ int _DkGetCPUInfo(PAL_CPU_INFO* ci) {
     }
 
     return rv;
+}
+
+#if USE_ARCH_RDRAND == 1
+int _DkRandomBitsRead(void* buffer, size_t size) {
+    uint32_t rand;
+    for (size_t i = 0; i < size; i += sizeof(rand)) {
+        rand = rdrand();
+        memcpy(buffer + i, &rand, MIN(sizeof(rand), size - i));
+    }
+    return 0;
+}
+#endif
+
+int _DkSegmentRegisterSet(int reg, const void* addr) {
+    int ret = 0;
+
+    if (reg == PAL_SEGMENT_FS) {
+        ret = INLINE_SYSCALL(arch_prctl, 2, ARCH_SET_FS, addr);
+    } else if (reg == PAL_SEGMENT_GS) {
+        return -PAL_ERROR_DENIED;
+    } else {
+        return -PAL_ERROR_INVAL;
+    }
+    if (IS_ERR(ret))
+        return -PAL_ERROR_DENIED;
+
+    return 0;
+}
+
+int _DkSegmentRegisterGet(int reg, void** addr) {
+    int ret;
+    unsigned long ret_addr;
+
+    if (reg == PAL_SEGMENT_FS) {
+        ret = INLINE_SYSCALL(arch_prctl, 2, ARCH_GET_FS, &ret_addr);
+    } else if (reg == PAL_SEGMENT_GS) {
+        // The GS segment is used for the internal TCB of PAL
+        return -PAL_ERROR_DENIED;
+    } else {
+        return -PAL_ERROR_INVAL;
+    }
+
+    if (IS_ERR(ret))
+        return -PAL_ERROR_DENIED;
+
+    *addr = (void*)ret_addr;
+    return 0;
+}
+
+int _DkCpuIdRetrieve(unsigned int leaf, unsigned int subleaf, unsigned int values[4]) {
+    cpuid(leaf, subleaf, values);
+    return 0;
 }
